Update constant variables and how they are used
Changed the variables containing OS-specific data (e.g constantLinuxDataSymbols) to be structs in order to group them with new/future OS-specific variables.
This also removes the need to create an identical function for the FreeBSD generator, which is similar to the one used for Linux profiles, but the difference
being a couple of variables.